backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
PyJWT==2.8.0
pydantic-settings==2.2.1
websockets==12.0